+++ src/main/java/frc/robot/testing/TestAutoCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.testing;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
