+++ src/main/java/frc/robot/unused/RainbowLedCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.unused;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
